Add Host.CreateDefaultBuilder #597
