+++ src/Hosting/Hosting/src/Internal/ServiceFactoryAdapter.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -9,19 +9,41 @@ namespace Microsoft.Extensions.Hosting.Internal
     internal class ServiceFactoryAdapter<TContainerBuilder> : IServiceFactoryAdapter
     {
         private IServiceProviderFactory<TContainerBuilder> _serviceProviderFactory;
+        private readonly Func<HostBuilderContext> _contextResolver;
+        private Func<HostBuilderContext, IServiceProviderFactory<TContainerBuilder>> _factoryResolver;
 
         public ServiceFactoryAdapter(IServiceProviderFactory<TContainerBuilder> serviceProviderFactory)
         {
-            _serviceProviderFactory = serviceProviderFactory ?? throw new System.ArgumentNullException(nameof(serviceProviderFactory));
+            _serviceProviderFactory = serviceProviderFactory ?? throw new ArgumentNullException(nameof(serviceProviderFactory));
+        }
+
+        public ServiceFactoryAdapter(Func<HostBuilderContext> contextResolver, Func<HostBuilderContext, IServiceProviderFactory<TContainerBuilder>> factoryResolver)
+        {
+            _contextResolver = contextResolver ?? throw new ArgumentNullException(nameof(contextResolver));
+            _factoryResolver = factoryResolver ?? throw new ArgumentNullException(nameof(factoryResolver));
         }
 
         public object CreateBuilder(IServiceCollection services)
         {
+            if (_serviceProviderFactory == null)
+            {
+                _serviceProviderFactory = _factoryResolver(_contextResolver());
+
+                if (_serviceProviderFactory == null)
+                {
+                    throw new InvalidOperationException("The resolver returned a null IServiceProviderFactory");
+                }
+            }
             return _serviceProviderFactory.CreateBuilder(services);
         }
 
         public IServiceProvider CreateServiceProvider(object containerBuilder)
         {
+            if (_serviceProviderFactory == null)
+            {
+                throw new InvalidOperationException("CreateBuilder must be called before CreateServiceProvider");
+            }
+
             return _serviceProviderFactory.CreateServiceProvider((TContainerBuilder)containerBuilder);
         }
     }
